Default to installing into sys.prefix instead of user-wide
This will work fine when using a virtual environment or some other
Python installation where the user has write permission, which is
relatively common.

The --user option is no longer the default and only suggested as a
fallback if the user has no permission to install into sys.prefix,
e. g. when using a system-wide Python installation directly.

There is no special support for installing into the system-wide location
anymore, but this can still be done using an explicit
--prefix=/usr/local or similar.